+++ lib/main.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:google_sign_in/google_sign_in.dart';
 
 import 'package:feetback/routes.dart';
 
@@ -22,9 +24,11 @@ class FeetBackApp extends StatelessWidget {
 
   static MaterialColor primaryColor = MaterialColor(0xFF33425B, primaryColorCodes);
 
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
+
     return MaterialApp(
       title: 'Feetback',
       theme: ThemeData(
@@ -33,7 +37,7 @@ class FeetBackApp extends StatelessWidget {
         accentColor: const Color(0xFFF33535),
         scaffoldBackgroundColor: const Color(0xFFD8E9F0),
       ),
-      initialRoute: "/",
+      initialRoute:"/",
       routes: routes,
       debugShowCheckedModeBanner: true,
     );
